requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
app==0.0.1
bcrypt==5.0.0
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.2
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
